+++ Spika/Data/Repository/Main/AppRepository+Sync.swift
@@ -1,3 +1,12 @@
+//
+//  AppRepository+Device.swift
+//  Spika
+//
+//  Created by Nikola Barbarić on 25.04.2022..
+//
+
+import Contacts
+import CoreTelephony
 import Combine
 import Foundation
 
@@ -185,6 +194,42 @@ extension AppRepository {
                 self?.saveUsers(users)
             })
             .store(in: &subs)
+    }
+    
+    func getPhoneContacts() -> Future<ContactFetchResult, Error> {
+        return Future() { promise in
+            let store = CNContactStore()
+            var contacts = [FetchedContact]()
+            CNContactStore().requestAccess(for: .contacts) { granted, error in
+                if let error = error {
+                    print("failed to request access", error)
+                    promise(.success(ContactFetchResult(error: error)))
+                    return
+                }
+                
+                if granted {
+                    let keys = [CNContactGivenNameKey, CNContactFamilyNameKey, CNContactPhoneNumbersKey]
+                    let request = CNContactFetchRequest(keysToFetch: keys as [CNKeyDescriptor])
+                    DispatchQueue.global(qos: .background).async {
+//                        let timer = ParkBenchTimer()
+                        do {
+                            try store.enumerateContacts(with: request, usingBlock: { (contact, stopPointer) in
+                                let fetchedContacts = self.phoneNumberParser.parse(contact.phoneNumbers.map { $0.value.stringValue })
+                                    .map { FetchedContact(firstName: contact.givenName, lastName: contact.familyName, telephone: $0) }
+                                
+                                contacts.append(contentsOf: fetchedContacts)
+                            })
+                        } catch let error {
+                            print("Failed to enumerate contact", error)
+                        }
+//                        print("Contact Pull finished: \(contacts.count), duration: \(timer.stop())")
+                        promise(.success(ContactFetchResult(fetchedContacts: contacts)))
+                    }
+                } else {
+                    promise(.success(ContactFetchResult()))
+                }
+            }
+        }
     }
 }
